Update main.js
Updated with messaging function and fix bug for the clicking on more button that caused request page to be opened when clicked on a class. Messaging page has a few bugs like the layout bug for the chat and incorrect username displayed on the sidechat. Suppose to display the sender's username instead of owners username.